+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All Notable changes to `speicher210/fastbill-bundle` will be documented in this 
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 2016-01-18
+## [1.1.1] 2016-02-04
+
+### Added
+- Command for listing customers
+- Command for resetting the customer
+- Command for listing subscriptions
+
+## [1.1.0] 2016-02-03
+
+### Added
+- Updated to latest API version
+
+## [1.0.0] 2016-01-18
 
 ### Added
 - Initial commit